New Updates

+++ TestSutraUsers/src/main/java/com/tathvatech/user/OID/Authorizable.java
@@ -2,11 +2,12 @@ package com.tathvatech.user.OID;
 
 import java.util.List;
 
-import com.tathvatech.common.EntityType;
+
+import com.tathvatech.common.enums.EntityType;
 
 public interface Authorizable 
 {
-	public int getPk();
+	public long getPk();
 	public EntityType getEntityType();
 	public String getDisplayText();
 	public List<? extends Role> getSupportedRoles();
