+++ TestSutraUsers/src/main/java/com/tathvatech/user/OID/ImmediateActionRequiredOID.java
@@ -2,7 +2,7 @@ package com.tathvatech.user.OID;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.tathvatech.common.EntityTypeEnum;
+import com.tathvatech.common.enums.EntityTypeEnum;
 
 public class ImmediateActionRequiredOID extends OID {
 
@@ -17,7 +17,7 @@ public class ImmediateActionRequiredOID extends OID {
 
 	@Override
 	public int hashCode() {
-		return getPk();
+		return (int) getPk();
 	}
 
 	@Override
