+++ TestSutraUsers/src/main/java/com/tathvatech/user/OID/LocationTypeOID.java
@@ -1,7 +1,7 @@
 package com.tathvatech.user.OID;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.tathvatech.common.EntityTypeEnum;
+import com.tathvatech.common.enums.EntityTypeEnum;
 
 
 public class LocationTypeOID extends OID{
@@ -43,7 +43,7 @@ public class LocationTypeOID extends OID{
 	
 	@Override
 	public int hashCode() {
-		return super.getPk();
+		return (int) super.getPk();
 	}
 
 }
