+++ TestSutraUsers/src/main/java/com/tathvatech/user/OID/MRFPartAssignOID.java
@@ -2,7 +2,8 @@ package com.tathvatech.user.OID;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.tathvatech.common.EntityTypeEnum;
+import com.tathvatech.common.enums.EntityTypeEnum;
+
 
 public class MRFPartAssignOID extends OID {
 
@@ -19,7 +20,7 @@ public class MRFPartAssignOID extends OID {
 
 	@Override
 	public int hashCode() {
-		return getPk();
+		return (int) getPk();
 	}
 
 	@Override
