+++ TestSutraUsers/src/main/java/com/tathvatech/user/OID/TSBeanBase.java
@@ -2,6 +2,6 @@ package com.tathvatech.user.OID;
 
 public abstract class TSBeanBase
 {
-	public abstract int getPk();
+	public abstract long getPk();
 	public abstract String getDisplayText();
 }
