+++ TestSutraUsers/src/main/java/com/tathvatech/user/OID/UnitOID.java
@@ -2,7 +2,7 @@ package com.tathvatech.user.OID;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.tathvatech.common.EntityTypeEnum;
+import com.tathvatech.common.enums.EntityTypeEnum;
 
 
 public class UnitOID extends OID{
@@ -45,7 +45,7 @@ public class UnitOID extends OID{
 	
 	@Override
 	public int hashCode() {
-		return super.getPk();
+		return (int) super.getPk();
 	}
 
 }
