+++ TestSutraUsers/src/main/java/com/tathvatech/user/OID/UserOID.java
@@ -3,7 +3,8 @@ package com.tathvatech.user.OID;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.tathvatech.common.EntityTypeEnum;
+import com.tathvatech.common.enums.EntityTypeEnum;
+import com.tathvatech.common.enums.EntityTypeEnum;
 
 
 @JsonIgnoreProperties(ignoreUnknown = true)
@@ -15,19 +16,19 @@ public class UserOID extends OID{
 	}
 	
 	@JsonCreator
-	public UserOID(@JsonProperty("pk") int pk)
+	public UserOID(@JsonProperty("pk") long pk)
 	{
 		super(pk, EntityTypeEnum.User, null);
 	}
 
-	public UserOID(int pk, String displayText)
+	public UserOID(long pk, String displayText)
 	{
 		super(pk, EntityTypeEnum.User, displayText);
 	}
 
 	@Override
 	public int hashCode() {
-		return getPk();
+		return (int) getPk();
 	}
 
 	@Override
