+++ TestSutraUsers/src/main/java/com/tathvatech/user/OID/WorkstationOID.java
@@ -2,7 +2,7 @@ package com.tathvatech.user.OID;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
-import com.tathvatech.common.EntityTypeEnum;
+import com.tathvatech.common.enums.EntityTypeEnum;
 
 
 public class WorkstationOID extends OID{
@@ -25,7 +25,7 @@ public class WorkstationOID extends OID{
 
 	@Override
 	public int hashCode() {
-		return getPk();
+		return (int) getPk();
 	}
 
 	@Override
