+++ TestSutraUsers/src/main/java/com/tathvatech/user/entity/SiteCache.java
@@ -6,7 +6,7 @@ import com.tathvatech.user.OID.SiteOID;
 import java.util.HashMap;
 
 
-public class SiteCache extends HashMap<Integer, Site> {
+public class SiteCache extends HashMap<Long, Site> {
 
 	private static SiteCache instance;
 	private SiteCache()
@@ -24,7 +24,7 @@ public class SiteCache extends HashMap<Integer, Site> {
 		Site sd = this.get(siteOID.getPk());
 		if(sd == null)
 		{
-			sd = (Site) new CommonServicesDelegate().getObjectByPk(Site.class, siteOID.getPk());
+			sd = (Site) new CommonServicesDelegate().getObjectByPk(Site.class, (int) siteOID.getPk());
 			this.put(siteOID.getPk(), sd);
 		}
 		return sd;
